scheduler/requirements.txt
azure-functions==1.20.0
requests==2.32.3
